services/auth/requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
passlib[bcrypt,argon2]
fastapi-jwt-auth
pydantic
pydantic[email]
bcrypt
argon2-cffi
pytest
pytest-asyncio
httpx
pydantic
pydantic-settings